+++ packages/calcite-components/src/tests/commonTests/t9n.ts
@@ -2,7 +2,8 @@ import { E2EElement, E2EPage } from "@stencil/core/testing";
 import { toHaveNoViolations } from "jest-axe";
 import { MessageBundle } from "../../utils/t9n";
 import { IntrinsicElementsWithProp } from "./../utils";
-import { getTagAndPage, ComponentTestSetup } from "./utils";
+import { getTagAndPage } from "./utils";
+import { ComponentTestSetup } from "./interfaces";
 
 expect.extend(toHaveNoViolations);
